+++ stalkerrender/source/base/SkeletonAnimated.cpp
@@ -81,7 +81,7 @@ static LPCSTR name_blend_type( CBlend::ECurvature blend )
 		{"eFREE_SLOT"		, CBlend::eFREE_SLOT	},
 		{"eAccrue"			, CBlend::eAccrue		},
 		{"eFalloff"			, CBlend::eFalloff		},
-		{"eFORCEDWORD"		, CBlend::eFORCEDWORD	}
+		{"eFORCEDWORD"		,static_cast<int>( CBlend::eFORCEDWORD	)}
 	};
 	return get_token_name( token_blend, blend );
 }
